Created udviewer.  An SDL project intended to replace udglViewer. udviewer runs on Windows x64, x86 and Linux.
Kernel::StreamerUpdate() now marks the scene as dirty if the streamer is active.
Added removeFirst and removeFirstSwapLast to udSlice.h.
Updated SDL driver systems to match with Qts.
Fixed a nasty bug Node::removeFirst() where NodeRef was implicitly casting to a bool.
Fixed a nasty bug in SDL Kernel::RunMainLoop() where the message queue was only processing one message per SDL_GL_SwapWindow() call.
Put an error path into SDLs Kernel::CreateInstanceInternal().
Kernel::DoInit() now returns an error code.
Created udRenderScene_Deinit() and udRenderScene_DeinitRener(). Will flesh these out when the destruction path is implemented.
Replaced udUIComponent_NUll.cpp with ui_null.cpp
Changed .gitmodules to not contain .euclideon.local so it works with linux vms.
Added error path to udRenderScene_InitRender().
Renderer::UDThread() now checks the return value of the call to udWaitSemaphore(pUDSemaphore).  Note this check should be added to all synchronisation function calls.
Fixed many many warnings.

+++ udKernel/src/components/datasources/geomsource.cpp
@@ -84,10 +84,10 @@ void GeomSource::Create(StreamRef spSource)
 
   // allocate for file
   int64_t len = spSource->Length();
-  void *pBuffer = udAlloc(len);
+  void *pBuffer = udAlloc((size_t)len);
 
   // read file from source
-  size_t read = spSource->Read(pBuffer, len);
+  size_t read = spSource->Read(pBuffer, (size_t)len);
   UDASSERT((int64_t)read == len, "!");
 
   Assimp::Importer importer;
